dataops-automation-ai/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
requests==2.32.3
pydantic==2.9.2
python-dotenv==1.0.1
APScheduler==3.10.4
streamlit==1.39.0
